+++ src/spring/azext_spring/tests/latest/test_asa_application_configuration_service.py
@@ -3,8 +3,6 @@
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
 
-import os
-
 from azure.cli.testsdk import (ScenarioTest, record_only)
 
 # pylint: disable=line-too-long
@@ -54,5 +52,11 @@ class ApplicationConfigurationServiceTest(ScenarioTest):
         self.cmd('spring application-configuration-service unbind --app {app} -g {rg} -s {serviceName}')
 
         self.cmd('spring application-configuration-service clear -g {rg} -s {serviceName}', checks=[
+            self.check('properties.provisioningState', "Succeeded")
+        ])
+
+        self.cmd('spring application-configuration-service delete -g {rg} -s {serviceName} --yes')
+
+        self.cmd('spring application-configuration-service create -g {rg} -s {serviceName}', checks=[
             self.check('properties.provisioningState', "Succeeded")
         ])